+++ source/extensions/omni.isaac.lab/omni/isaac/lab/envs/manager_based_env_cfg.py
@@ -101,8 +101,9 @@ class ManagerBasedEnvCfg:
     Defaults to False, which means no render step will be performed after reset.
 
     * When this is False, data collected from sensors after performing reset will be stale and will not reflect the
-    latest states in simulation caused by the reset.
+      latest states in simulation caused by the reset.
     * When this is True, an extra render step will be performed to update the sensor data
-    to reflect the latest states from the reset. This comes at a cost of performance as an additional render
-    step will be performed after each time an environment is reset.
+      to reflect the latest states from the reset. This comes at a cost of performance as an additional render
+      step will be performed after each time an environment is reset.
+
     """