+++ source/standalone/workflows/skrl/play.py
@@ -109,7 +109,9 @@ def main():
     if args_cli.checkpoint:
         resume_path = os.path.abspath(args_cli.checkpoint)
     else:
-        resume_path = get_checkpoint_path(log_root_path, other_dirs=["checkpoints"])
+        resume_path = get_checkpoint_path(
+            log_root_path, run_dir=f".*_{algorithm}_{args_cli.ml_framework}", other_dirs=["checkpoints"]
+        )
     log_dir = os.path.dirname(os.path.dirname(resume_path))
 
     # create isaac environment